refactor(web): reorganise actions to queries
