--- apps/web/src/actions/cars/index.ts
@@ -1,1 +1,0 @@
-export { getPopularMakes } from "./popular-makes";
