--- apps/web/src/actions/coe/index.ts
@@ -1,1 +1,0 @@
-export * from "./pqp";
