+++ apps/web/src/app/(dashboard)/(home)/page.tsx
@@ -1,7 +1,3 @@
-import {
-  getTopMakesByYear,
-  getYearlyRegistrations,
-} from "@web/actions/cars/statistics";
 import { KeyStatistics } from "@web/app/(dashboard)/(home)/_components/key-statistics";
 import { RecentPosts } from "@web/app/(dashboard)/(home)/_components/recent-posts";
 import { LatestCOE } from "@web/components/coe/latest-coe";
@@ -9,8 +5,9 @@ import { StructuredData } from "@web/components/structured-data";
 import { TopMakesByYear } from "@web/components/top-makes-by-year";
 import { TotalNewCarRegistrationsByYear } from "@web/components/total-new-car-registrations-by-year";
 import { SITE_TITLE, SITE_URL } from "@web/config";
-import { getLatestCOEResults } from "@web/lib/coe/queries";
 import { getAllPosts } from "@web/lib/data/posts";
+import { getTopMakesByYear, getYearlyRegistrations } from "@web/queries/cars";
+import { getLatestCOEResults } from "@web/queries/coe";
 import type { Metadata } from "next";
 import type { WebSite, WithContext } from "schema-dts";
