+++ apps/web/src/app/(dashboard)/cars/[category]/category-tabs.tsx
@@ -11,7 +11,7 @@ import {
 import type {
   CarMarketShareResponse,
   CarTopPerformersData,
-} from "@web/lib/cars/queries";
+} from "@web/queries/cars";
 import type { TypeItem } from "@web/types";
 import { formatPercentage } from "@web/utils/charts";
 import { formatDateToMonthYear } from "@web/utils/format-date-to-month-year";
