+++ apps/web/src/app/(dashboard)/cars/[category]/page.tsx
@@ -4,12 +4,12 @@ import { PageHeader } from "@web/components/page-header";
 import { StructuredData } from "@web/components/structured-data";
 import Typography from "@web/components/typography";
 import { LAST_UPDATED_CARS_KEY, SITE_TITLE, SITE_URL } from "@web/config";
+import { createPageMetadata } from "@web/lib/metadata";
 import {
   getCarMarketShareData,
   getCarsData,
   getCarTopPerformersData,
-} from "@web/lib/cars/queries";
-import { createPageMetadata } from "@web/lib/metadata";
+} from "@web/queries/cars";
 import { formatDateToMonthYear } from "@web/utils/format-date-to-month-year";
 import { fetchMonthsForCars, getMonthOrLatest } from "@web/utils/months";
 import type { Metadata } from "next";
