+++ apps/web/src/app/(dashboard)/cars/makes/page.tsx
@@ -1,14 +1,13 @@
 "use cache";
 
 import { redis } from "@sgcarstrends/utils";
-import { getPopularMakes } from "@web/actions";
-import { getAllCarLogos } from "@web/actions/logos";
 import { MakesList } from "@web/app/(dashboard)/cars/_components/makes";
 import { PageHeader } from "@web/components/page-header";
 import { StructuredData } from "@web/components/structured-data";
 import { LAST_UPDATED_CARS_KEY, SITE_TITLE, SITE_URL } from "@web/config";
-import { getDistinctMakes } from "@web/lib/cars/queries";
 import { createPageMetadata } from "@web/lib/metadata";
+import { getDistinctMakes, getPopularMakes } from "@web/queries/cars";
+import { getAllCarLogos } from "@web/queries/logos";
 import { fetchMonthsForCars } from "@web/utils/months";
 import type { Metadata } from "next";
 import type { WebPage, WithContext } from "schema-dts";
