+++ apps/web/src/app/(dashboard)/cars/page.tsx
@@ -9,14 +9,14 @@ import { StructuredData } from "@web/components/structured-data";
 import Typography from "@web/components/typography";
 import { UnreleasedFeature } from "@web/components/unreleased-feature";
 import { LAST_UPDATED_CARS_KEY, SITE_TITLE, SITE_URL } from "@web/config";
+import { createPageMetadata } from "@web/lib/metadata";
+import { generateDatasetSchema } from "@web/lib/structured-data";
 import {
   getCarsComparison,
   getCarsData,
   getTopMakesByFuelType,
   getTopTypes,
-} from "@web/lib/cars/queries";
-import { createPageMetadata } from "@web/lib/metadata";
-import { generateDatasetSchema } from "@web/lib/structured-data";
+} from "@web/queries/cars";
 import { formatDateToMonthYear } from "@web/utils/format-date-to-month-year";
 import { fetchMonthsForCars, getMonthOrLatest } from "@web/utils/months";
 import type { Metadata } from "next";
