+++ apps/web/src/app/(dashboard)/coe/latest/page.tsx
@@ -1,8 +1,8 @@
 import { redis } from "@sgcarstrends/utils";
 import { StructuredData } from "@web/components/structured-data";
 import { LAST_UPDATED_COE_KEY, SITE_TITLE, SITE_URL } from "@web/config";
-import { getLatestCOEResults } from "@web/lib/coe/queries";
 import { createPageMetadata } from "@web/lib/metadata";
+import { getLatestCOEResults } from "@web/queries/coe";
 import type { Metadata } from "next";
 import type { WebPage, WithContext } from "schema-dts";
 import { LatestCOEClient } from "./latest-coe-content";
