+++ apps/web/src/app/(dashboard)/coe/page.tsx
@@ -16,12 +16,12 @@ import { StructuredData } from "@web/components/structured-data";
 import Typography from "@web/components/typography";
 import { LAST_UPDATED_COE_KEY, SITE_TITLE, SITE_URL } from "@web/config";
 import { calculateOverviewStats } from "@web/lib/coe/calculations";
+import { createPageMetadata } from "@web/lib/metadata";
 import {
   getCOEResults,
   getLatestCOEResults,
-  getPQPData,
-} from "@web/lib/coe/queries";
-import { createPageMetadata } from "@web/lib/metadata";
+  getPqpRates,
+} from "@web/queries/coe";
 import { formatPercent } from "@web/utils/charts";
 import { formatDateToMonthYear } from "@web/utils/format-date-to-month-year";
 import type { Metadata } from "next";
@@ -57,7 +57,7 @@ const COEPricesPage = async () => {
   const [latestResults, allCoeResults, pqpRates] = await Promise.all([
     getLatestCOEResults(),
     getCOEResults(),
-    getPQPData(),
+    getPqpRates(),
   ]);
   const lastUpdated = await redis.get<number>(LAST_UPDATED_COE_KEY);
