+++ apps/web/src/app/sitemap.ts
@@ -1,7 +1,7 @@
 import slugify from "@sindresorhus/slugify";
 import { SITE_LINKS, SITE_URL } from "@web/config";
-import { getDistinctMakes } from "@web/lib/cars/queries";
 import { getAllPosts } from "@web/lib/data/posts";
+import { getDistinctMakes } from "@web/queries/cars";
 import type { MetadataRoute } from "next";
 
 const sitemap = async (): Promise<MetadataRoute.Sitemap> => {
