+++ apps/web/src/queries/coe/available-months.ts
@@ -1,0 +1,16 @@
+import { coe, db } from "@sgcarstrends/database";
+import { desc } from "drizzle-orm";
+import { cacheLife, cacheTag } from "next/cache";
+
+export const getCoeMonths = async (): Promise<{ month: string }[]> => {
+  "use cache";
+  cacheLife("statistics");
+  cacheTag("coe", "coe-months");
+
+  const results = await db
+    .selectDistinct({ month: coe.month })
+    .from(coe)
+    .orderBy(desc(coe.month));
+
+  return results.map((r) => ({ month: r.month ?? "" }));
+};
