+++ apps/web/src/queries/coe/index.ts
@@ -1,0 +1,4 @@
+export * from "./available-months";
+export * from "./historical-results";
+export * from "./latest-results";
+export * from "./pqp";
