+++ apps/web/src/queries/coe/pqp/overview.ts
@@ -1,5 +1,3 @@
-"use server";
-
 import { coe, coePQP, db } from "@sgcarstrends/database";
 import type { Pqp } from "@web/types/coe";
 import { and, desc, eq, inArray, isNotNull } from "drizzle-orm";
