+++ apps/web/src/queries/index.ts
@@ -1,0 +1,3 @@
+export * from "./cars";
+export * from "./coe";
+export * from "./logos";
